requirements.txt
Flask
requests
beautifulsoup4
lxml
discord
python-dotenv